requirements.txt
setuptools==70.0.0  # temp fix for compatibility with some old packages
GitPython==3.1.32
Pillow==10.3.0
accelerate==0.21.0
blendmodes==2023
clean-fid==0.1.35
diskcache==5.6.3
einops==0.4.1
facexlib==0.3.0
fastapi>=0.109.1
gradio~=5.6.0
httpcore==0.15
inflection==0.5.1
jsonmerge==1.8.0
kornia==0.6.7
lark==1.1.2
numpy==1.26.2
omegaconf==2.2.3
open-clip-torch~=2.29.0
piexif==1.1.3
protobuf>=4.21.6
psutil==5.9.5
pytorch_lightning==1.9.4
resize-right==0.0.2
safetensors==0.4.2
scikit-image==0.21.0
spandrel==0.3.4
spandrel-extra-arches==0.1.1
tomesd==0.1.3
torch
torchdiffeq==0.2.3
torchsde==0.2.6
transformers==4.44.0
httpx==0.24.1
pillow-avif-plugin==1.4.3
diffusers==0.29.2
gradio_rangeslider~=0.0.6
gradio_imageslider~=0.0.20
loadimg==0.1.2
tqdm==4.66.3
peft==0.12.0
pydantic==2.8.2
huggingface-hub==0.25.1
python-dotenv~=1.0.1
streamlit~=1.40.1
requests~=2.32.3
imageio~=2.36.0
together~=1.3.5